chore: create user test data with util method
