+++ packages/db/test/user-storage.spec.js
@@ -2,11 +2,8 @@
 import assert from 'assert'
 import { DBClient } from '../index.js'
 import {
-  createUser,
-  createUserAuthKey,
-  createUpload,
-  createPsaPinRequest,
-  token
+  token,
+  createUserWithFiles
 } from './utils.js'
 
 describe('Users used storage', () => {
@@ -17,113 +14,27 @@ describe('Users used storage', () => {
     postgres: true
   })
 
-  let user1, user2, user3, user4
-  let authKey1, authKey2, authKey3, authKey4
-  const cids = [
-    'bafybeiczsscdsbs7ffqz55asqdf3smv6klcw3gofszvwlyarci47fgf111',
-    'bafybeiczsscdsbs7ffqz55asqdf3smv6klcw3gofszvwlyarci47fgf112',
-    'bafybeiczsscdsbs7ffqz55asqdf3smv6klcw3gofszvwlyarci47fgf113'
-  ]
-  const cidsPinned = [
-    'QmdA5WkDNALetBn4iFeSepHjdLGJdxPBwZyY47ir1bZGAK',
-    'QmNvTjdqEPjZVWCvRWsFJA1vK7TTw1g9JP6we1WBJTRADM'
-  ]
-  const uploadSize = 439804651110
-  const pinnedSize = 109951162800
-  const largeFileSize = 3199023255550
-  const pins = [
-    {
-      status: 'Pinning',
-      location: {
-        peerId: '12D3KooWFe387JFDpgNEVCP5ARut7gRkX7YuJCXMStpkq714ziK6',
-        peerName: 'web3-storage-sv15',
-        region: 'region'
-      }
-    },
-    {
-      status: 'Pinned',
-      location: {
-        peerId: '12D3KooWFe387JFDpgNEVCP5ARut7gRkX7YuJCXMStpkq714ziK7',
-        peerName: 'web3-storage-sv16',
-        region: 'region'
-      }
-    }
-  ]
-
   beforeEach(async () => {
-    user1 = await createUser(dbClient, {
-      name: 'test1-name',
-      email: 'test1@email.com'
-    })
-    authKey1 = await createUserAuthKey(dbClient, Number(user1._id), {
-      name: 'test1-key'
-    })
-    await createUpload(dbClient, Number(user1._id), Number(authKey1), cids[0], {
-      dagSize: uploadSize
-    })
-    await createPsaPinRequest(dbClient, authKey1, cidsPinned[0], {
-      dagSize: pinnedSize,
-      pins
-    })
-    await createPsaPinRequest(dbClient, authKey1, cidsPinned[1], {
-      dagSize: pinnedSize,
-      pins
+    await createUserWithFiles(dbClient, {
+      email: 'test1@email.com',
+      percentStorageUsed: 60
     })
 
-    user2 = await createUser(dbClient, {
-      name: 'test2-name',
-      email: 'test2@email.com'
-    })
-    authKey2 = await createUserAuthKey(dbClient, Number(user2._id), {
-      name: 'test2-key'
-    })
-    await createUpload(dbClient, Number(user2._id), Number(authKey2), cids[0], {
-      dagSize: uploadSize
-    })
-    await createUpload(dbClient, Number(user2._id), Number(authKey2), cids[1], {
-      dagSize: uploadSize
+    await createUserWithFiles(dbClient, {
+      email: 'test2@email.com',
+      percentStorageUsed: 79
     })
 
-    user3 = await createUser(dbClient, {
-      name: 'test3-name',
-      email: 'test3@email.com'
-    })
-    authKey3 = await createUserAuthKey(dbClient, Number(user3._id), {
-      name: 'test3-key'
-    })
-    await createUpload(dbClient, Number(user3._id), Number(authKey3), cids[0], {
-      dagSize: uploadSize
-    })
-    await createUpload(dbClient, Number(user3._id), Number(authKey3), cids[1], {
-      dagSize: Math.round(uploadSize * 1.2)
-    })
-    await createPsaPinRequest(dbClient, authKey3, cidsPinned[0], {
-      dagSize: pinnedSize,
-      pins
+    await createUserWithFiles(dbClient, {
+      email: 'test3@email.com',
+      percentStorageUsed: 90
     })
 
-    user4 = await createUser(dbClient, {
-      name: 'test4-name',
-      email: 'test4@email.com'
+    await createUserWithFiles(dbClient, {
+      email: 'test4@email.com',
+      percentStorageUsed: 145,
+      storageQuota: 2199023255552
     })
-    authKey4 = await createUserAuthKey(dbClient, Number(user4._id), {
-      name: 'test4-key'
-    })
-    await dbClient.createUserTag(Number(user4._id), {
-      tag: 'StorageLimitBytes',
-      value: '2199023255552'
-    })
-    const cid = 'bafybeibvuy3vcepqxy4plr34twv22vvxol2jjhmjxcrcvuhea5226whpsm'
-    await createUpload(dbClient, Number(user4._id), Number(authKey4), cid, {
-      dagSize: largeFileSize
-    })
-
-    // User               | upload  | pinned  | quota   | percentage of quota
-    // ----------------------------------------------------------------------
-    // test1@email.com    | 1       | 2       | 1TiB    | < 75%
-    // test2@email.com    | 2       | 0       | 1TiB    | 80%
-    // test3@email.com    | 2       | 1       | 1TiB    | 90%
-    // test4@email.com    | 1 (XL)  | 0       | 2TiB    | > 90%
   })
 
   it('returns user details needed for email', async () => {
